re-organizing for a hopefully cleaner PR
